+++ activities/MultipliersAndDivisors/CHANGELOG.md
@@ -1,3 +1,4 @@
 ## HEAD
 
-- 🆕 [v1.0.0] `openThePortalProcessorBuilder`: Process the `OPEN_PORTAL` instruction.
+- 🚀 [v1.0.0] `openThePortalProcessorBuilder`: Process the `OPEN_PORTAL` instruction.
+- 🚀 [v1.0.0] `hasLeftTheMazeProcessorBuilder`: Process the `LEAVE_MAZE` instruction, and it verifies if the number has entered the numeric lineand in which empty slot.